\name{itext}
\alias{itext}
\title{
Add text objects to an interactive plot
}
\description{
\code{itext} method adds text objects to an interactive plot.
}
\usage{
itext(x, ...)
}
\arguments{
  \item{x}{objects representing the text}
  \item{\dots}{further optional arguments}
}
%\details{
%}
\value{
  undefined
}
\seealso{
\code{\link{ilines}}, \code{\link{ipoints}}
}
%\examples{
%}
\keyword{iplot}
